Again the parameter reference
It seems the parameter reference only work with the 'parameters' prefix.
The reason of this behaviour is that the variables are to be resloved
when the artifact object is created. No further SPEL expression evalutation
happens the the later stages

test case

Create a pipeline define a parameter and a key in the 'Evaluate Variable section'.
Observe the differences